refactor old rag/judge.py
